print string

+++ 0x05-pointers_arrays_strings/3-puts.c
@@ -11,7 +11,7 @@
 
 void _puts(char *str)
 {
-	int i;
+	long unsigned int i;
 
 	for (i = 0; i <= strlen(str); i++)
 	{
